refactor: consolidate old functionality
Removed functionality no longer relevant to project and consolidated functionality into more logical file/folder structure and naming convention
